+++ sdk/tables/azure-data-tables/tests/test_table_batch_cosmos_async.py
@@ -52,7 +52,7 @@ class StorageTableBatchTest(AzureTestCase, AsyncTableTestCase):
                 await self.ts.create_table(self.table_name)
             except ResourceExistsError:
                 pass
-        self.test_tables = []
+        self.test_tables = [self.table_name]
 
     async def _tear_down(self):
         if self.is_live:
@@ -66,8 +66,9 @@ class StorageTableBatchTest(AzureTestCase, AsyncTableTestCase):
                     await self.ts.delete_table(table_name)
                 except:
                     pass
-            sleep(SLEEP_DELAY)
         await self.table.close()
+        self.test_tables = []
+        self.sleep(SLEEP_DELAY)
 
     #--Helpers-----------------------------------------------------------------
 
@@ -173,7 +174,6 @@ class StorageTableBatchTest(AzureTestCase, AsyncTableTestCase):
         assert length ==  len(transaction.requests)
 
     #--Test cases for batch ---------------------------------------------
-    @pytest.mark.skipif(sys.version_info < (3, 0), reason="requires Python3")
     @CosmosPreparer()
     async def test_batch_single_insert(self, tables_cosmos_account_name, tables_primary_cosmos_account_key):
         # Arrange
@@ -196,24 +196,21 @@ class StorageTableBatchTest(AzureTestCase, AsyncTableTestCase):
             # Assert
             self._assert_valid_batch_transaction(transaction_result, 1)
             sent_entity = transaction_result.get_entity(entity.RowKey)
+
             assert sent_entity is not None
-
             e = await self.table.get_entity(row_key=entity.RowKey, partition_key=entity.PartitionKey)
-
             assert e.test ==  entity.test.value
             assert e.test2 ==  entity.test2
             assert e.test3 ==  entity.test3
             assert e.test4 ==  entity.test4.value
-
             assert sent_entity['test'] ==  entity.test.value
             assert sent_entity['test2'] ==  entity.test2
             assert sent_entity['test3'] ==  entity.test3
             assert sent_entity['test4'] ==  entity.test4.value
+
         finally:
             await self._tear_down()
 
-    @pytest.mark.skip("merge operations fail in cosmos: https://github.com/Azure/azure-sdk-for-python/issues/13844")
-    @pytest.mark.skipif(sys.version_info < (3, 0), reason="requires Python3")
     @CosmosPreparer()
     async def test_batch_single_update(self, tables_cosmos_account_name, tables_primary_cosmos_account_key):
         # Arrange
@@ -245,12 +242,10 @@ class StorageTableBatchTest(AzureTestCase, AsyncTableTestCase):
             result = await self.table.get_entity(row_key=entity.RowKey, partition_key=entity.PartitionKey)
             assert result.PartitionKey ==  u'001'
             assert result.RowKey ==  u'batch_insert'
-            assert result.test3.value ==  5
+            assert result.test3 ==  5
         finally:
             await self._tear_down()
 
-    @pytest.mark.skip("merge operations fail in cosmos: https://github.com/Azure/azure-sdk-for-python/issues/13844")
-    @pytest.mark.skipif(sys.version_info < (3, 0), reason="requires Python3")
     @CosmosPreparer()
     async def test_batch_update(self, tables_cosmos_account_name, tables_primary_cosmos_account_key):
         # Arrange
@@ -268,7 +263,7 @@ class StorageTableBatchTest(AzureTestCase, AsyncTableTestCase):
             await self.table.create_entity(entity)
 
             entity = await self.table.get_entity(u'001', u'batch_update')
-            assert 3 ==  entity.test3.value
+            assert 3 ==  entity.test3
             entity.test2 = u'value1'
 
             batch = self.table.create_batch()
@@ -279,14 +274,12 @@ class StorageTableBatchTest(AzureTestCase, AsyncTableTestCase):
             self._assert_valid_batch_transaction(transaction_result, 1)
             assert transaction_result.get_entity(entity.RowKey) is not None
             result = await self.table.get_entity('001', 'batch_update')
-            assert 'value1' ==  result.test2.value
+            assert 'value1' ==  result.test2
             assert entity.PartitionKey ==  u'001'
             assert entity.RowKey ==  u'batch_update'
         finally:
             await self._tear_down()
 
-    @pytest.mark.skip("merge operations fail in cosmos: https://github.com/Azure/azure-sdk-for-python/issues/13844")
-    @pytest.mark.skipif(sys.version_info < (3, 0), reason="requires Python3")
     @CosmosPreparer()
     async def test_batch_merge(self, tables_cosmos_account_name, tables_primary_cosmos_account_key):
         # Arrange
@@ -319,14 +312,13 @@ class StorageTableBatchTest(AzureTestCase, AsyncTableTestCase):
             assert transaction_result.get_entity(entity.RowKey) is not None
 
             resp_entity = await self.table.get_entity(partition_key=u'001', row_key=u'batch_merge')
-            assert entity.test2 ==  resp_entity.test2.value
-            assert 1234567890 ==  resp_entity.test4.value
+            assert entity.test2 ==  resp_entity.test2
+            assert 1234567890 ==  resp_entity.test4
             assert entity.PartitionKey ==  resp_entity.PartitionKey
             assert entity.RowKey ==  resp_entity.RowKey
         finally:
             await self._tear_down()
 
-    @pytest.mark.skipif(sys.version_info < (3, 0), reason="requires Python3")
     @CosmosPreparer()
     async def test_batch_update_if_match(self, tables_cosmos_account_name, tables_primary_cosmos_account_key):
         # Arrange
@@ -356,7 +348,6 @@ class StorageTableBatchTest(AzureTestCase, AsyncTableTestCase):
         finally:
             await self._tear_down()
 
-    @pytest.mark.skipif(sys.version_info < (3, 0), reason="requires Python3")
     @CosmosPreparer()
     async def test_batch_update_if_doesnt_match(self, tables_cosmos_account_name, tables_primary_cosmos_account_key):
         # Arrange
@@ -385,8 +376,6 @@ class StorageTableBatchTest(AzureTestCase, AsyncTableTestCase):
         finally:
             await self._tear_down()
 
-    @pytest.mark.skip("merge operations fail in cosmos: https://github.com/Azure/azure-sdk-for-python/issues/13844")
-    @pytest.mark.skipif(sys.version_info < (3, 0), reason="requires Python3")
     @CosmosPreparer()
     async def test_batch_insert_replace(self, tables_cosmos_account_name, tables_primary_cosmos_account_key):
         # Arrange
@@ -412,13 +401,11 @@ class StorageTableBatchTest(AzureTestCase, AsyncTableTestCase):
 
             entity = await self.table.get_entity('001', 'batch_insert_replace')
             assert entity is not None
-            assert 'value' ==  entity.test2.value
-            assert 1234567890 ==  entity.test4.value
+            assert 'value' ==  entity.test2
+            assert 1234567890 ==  entity.test4
         finally:
             await self._tear_down()
 
-    @pytest.mark.skip("merge operations fail in cosmos: https://github.com/Azure/azure-sdk-for-python/issues/13844")
-    @pytest.mark.skipif(sys.version_info < (3, 0), reason="requires Python3")
     @CosmosPreparer()
     async def test_batch_insert_merge(self, tables_cosmos_account_name, tables_primary_cosmos_account_key):
         # Arrange
@@ -443,12 +430,11 @@ class StorageTableBatchTest(AzureTestCase, AsyncTableTestCase):
             assert transaction_result.get_entity(entity.RowKey) is not None
             entity = await self.table.get_entity('001', 'batch_insert_merge')
             assert entity is not None
-            assert 'value' ==  entity.test2.value
-            assert 1234567890 ==  entity.test4.value
+            assert 'value' ==  entity.test2
+            assert 1234567890 ==  entity.test4
         finally:
             await self._tear_down()
 
-    @pytest.mark.skipif(sys.version_info < (3, 0), reason="requires Python3")
     @CosmosPreparer()
     async def test_batch_delete(self, tables_cosmos_account_name, tables_primary_cosmos_account_key):
         # Arrange
@@ -481,7 +467,6 @@ class StorageTableBatchTest(AzureTestCase, AsyncTableTestCase):
         finally:
             await self._tear_down()
 
-    @pytest.mark.skipif(sys.version_info < (3, 0), reason="requires Python3")
     @CosmosPreparer()
     async def test_batch_inserts(self, tables_cosmos_account_name, tables_primary_cosmos_account_key):
         # Arrange
@@ -497,7 +482,7 @@ class StorageTableBatchTest(AzureTestCase, AsyncTableTestCase):
             transaction_count = 0
 
             batch = self.table.create_batch()
-            for i in range(20):
+            for i in range(10):
                 entity.RowKey = str(i)
                 batch.create_entity(entity)
                 transaction_count += 1
@@ -515,12 +500,10 @@ class StorageTableBatchTest(AzureTestCase, AsyncTableTestCase):
 
             # Assert
             assert entities is not None
-            assert transaction_count ==  length
+            assert 10 == length
         finally:
             await self._tear_down()
 
-    @pytest.mark.skip("merge operations fail in cosmos: https://github.com/Azure/azure-sdk-for-python/issues/13844")
-    @pytest.mark.skipif(sys.version_info < (3, 0), reason="requires Python3")
     @CosmosPreparer()
     async def test_batch_all_operations_together(self, tables_cosmos_account_name, tables_primary_cosmos_account_key):
         # Arrange
@@ -592,8 +575,6 @@ class StorageTableBatchTest(AzureTestCase, AsyncTableTestCase):
         finally:
             await self._tear_down()
 
-    @pytest.mark.skip("merge operations fail in cosmos: https://github.com/Azure/azure-sdk-for-python/issues/13844")
-    @pytest.mark.skipif(sys.version_info < (3, 0), reason="requires Python3")
     @CosmosPreparer()
     async def test_batch_all_operations_together_context_manager(self, tables_cosmos_account_name, tables_primary_cosmos_account_key):
         # Arrange
@@ -642,7 +623,7 @@ class StorageTableBatchTest(AzureTestCase, AsyncTableTestCase):
         finally:
             await self._tear_down()
 
-    @pytest.mark.skipif(sys.version_info < (3, 0), reason="requires Python3")
+    @pytest.mark.skip("The same row operations do not fail on Cosmos")
     @CosmosPreparer()
     async def test_batch_same_row_operations_fail(self, tables_cosmos_account_name, tables_primary_cosmos_account_key):
         # Arrange
@@ -657,6 +638,7 @@ class StorageTableBatchTest(AzureTestCase, AsyncTableTestCase):
             entity = self._create_updated_entity_dict(
                 '001', 'batch_negative_1')
             batch.update_entity(entity)
+
             entity = self._create_random_entity_dict(
                 '001', 'batch_negative_1')
             batch.update_entity(entity)
@@ -667,7 +649,6 @@ class StorageTableBatchTest(AzureTestCase, AsyncTableTestCase):
         finally:
             await self._tear_down()
 
-    @pytest.mark.skipif(sys.version_info < (3, 0), reason="requires Python3")
     @CosmosPreparer()
     async def test_batch_different_partition_operations_fail(self, tables_cosmos_account_name, tables_primary_cosmos_account_key):
         # Arrange
@@ -709,12 +690,12 @@ class StorageTableBatchTest(AzureTestCase, AsyncTableTestCase):
                     entity.PartitionKey = 'large'
                     entity.RowKey = 'item{0}'.format(i)
                     batch.create_entity(entity)
+                await self.table.send_batch(batch)
 
             # Assert
         finally:
             await self._tear_down()
 
-    @pytest.mark.skipif(sys.version_info < (3, 0), reason="requires Python3")
     @CosmosPreparer()
     async def test_new_non_existent_table(self, tables_cosmos_account_name, tables_primary_cosmos_account_key):
         # Arrange
@@ -733,7 +714,28 @@ class StorageTableBatchTest(AzureTestCase, AsyncTableTestCase):
         finally:
             await self._tear_down()
 
-    @pytest.mark.skipif(sys.version_info < (3, 0), reason="requires Python3")
+    @pytest.mark.live_test_only
+    @CosmosPreparer()
+    async def test_new_invalid_key(self, tables_cosmos_account_name, tables_primary_cosmos_account_key):
+        # Arrange
+        invalid_key = tables_primary_cosmos_account_key[0:-6] + "==" # cut off a bit from the end to invalidate
+        key_list = list(tables_primary_cosmos_account_key)
+
+        key_list[-6:] = list("0000==")
+        invalid_key = ''.join(key_list)
+
+        self.ts = TableServiceClient(self.account_url(tables_cosmos_account_name, "table"), invalid_key)
+        self.table_name = self.get_resource_name('uttable')
+        self.table = self.ts.get_table_client(self.table_name)
+
+        entity = self._create_random_entity_dict('001', 'batch_negative_1')
+
+        batch = self.table.create_batch()
+        batch.create_entity(entity)
+
+        with pytest.raises(ClientAuthenticationError):
+            resp = await self.table.send_batch(batch)
+
     @CosmosPreparer()
     async def test_new_delete_nonexistent_entity(self, tables_cosmos_account_name, tables_primary_cosmos_account_key):
         # Arrange